Configure keycloak basic auth for all pulp apis
Issue: AAH-916
