+++ CHANGES/916.feature
@@ -1,0 +1,1 @@
+Enable keycloak authentication using username and password for podman login.
