+++ dev/standalone-keycloak/galaxy_ng.env
@@ -18,7 +18,7 @@ PULP_CONTENT_ORIGIN="http://localhost:24816"
 # https://docs.pulpproject.org/pulp_container/workflows/authentication.html
 
 PULP_TOKEN_AUTH_DISABLED=false
-PULP_TOKEN_SERVER=http://localhost:5001/keycloak-container-token/
+PULP_TOKEN_SERVER=http://localhost:5001/token/
 PULP_TOKEN_SIGNATURE_ALGORITHM=ES256
 PULP_PUBLIC_KEY_PATH=/src/galaxy_ng/dev/common/container_auth_public_key.pem
 PULP_PRIVATE_KEY_PATH=/src/galaxy_ng/dev/common/container_auth_private_key.pem
