+++ galaxy_ng/app/dynaconf_hooks.py
@@ -114,6 +114,11 @@ def configure_keycloak(settings: Dynaconf) -> Dict[str, Any]:
             "dynaconf_merge",
         ]
 
+        data["REST_FRAMEWORK__DEFAULT_AUTHENTICATION_CLASSES"] = [
+            "rest_framework.authentication.SessionAuthentication",
+            "galaxy_ng.app.auth.keycloak.KeycloakBasicAuth",
+        ]
+
         # Replace AUTH CLASSES
         data["GALAXY_AUTHENTICATION_CLASSES"] = [
             "rest_framework.authentication.SessionAuthentication",
