+++ galaxy_ng/app/views.py
@@ -1,71 +1,5 @@
-import jwt
-from requests import post as requests_post
-
 from django.http import HttpResponse
-from django.conf import settings
-
-from rest_framework.authentication import BasicAuthentication
-
-from rest_framework import status as http_code
-from rest_framework import exceptions
-
-from social_django.utils import load_strategy
-from social_core.backends.keycloak import KeycloakOAuth2
-
-from gettext import gettext as _
-
-from pulp_container.app.registry_api import BearerTokenView
 
 
 def health_view(request):
     return HttpResponse('OK')
-
-
-class KeycloakBasicAuth(BasicAuthentication):
-    def authenticate_credentials(self, userid, password, request=None):
-        payload = {
-            'client_id': settings.SOCIAL_AUTH_KEYCLOAK_KEY,
-            'client_secret': settings.SOCIAL_AUTH_KEYCLOAK_SECRET,
-            'grant_type': 'password',
-            'scope': 'openid',
-            'username': userid,
-            'password': password
-        }
-        headers = {
-            "Content-Type": "application/x-www-form-urlencoded",
-        }
-
-        response = requests_post(
-            url=settings.SOCIAL_AUTH_KEYCLOAK_ACCESS_TOKEN_URL,
-            headers=headers,
-            data=payload
-        )
-
-        if response.status_code == http_code.HTTP_200_OK:
-
-            # load social auth django strategy
-            strategy = load_strategy(request)
-            backend = KeycloakOAuth2(strategy)
-
-            # TODO: How do you verify JWT signature and audience? 
-            token_data = jwt.decode(response.json()['access_token'], options={
-                "verify_signature": False,
-                "verify_aud": False
-            })
-
-            # The django social auth strategy uses data from the JWT token in the KeycloackOAuth2
-            # backend to create a new user and update it with the data from the token. This
-            # should return a django user instance.
-            user = strategy.authenticate(backend, response=token_data)
-
-            if user is None:
-                raise exceptions.AuthenticationFailed(_("Authentication failed."))
-
-            return (user, None)
-
-        else:
-            raise exceptions.AuthenticationFailed(_("Invalid username and password."))
-
-
-class ContainerRegistryTokenView(BearerTokenView):
-    authentication_classes = [KeycloakBasicAuth, ]
